reduce memory usage by using float

+++ helper.c
@@ -11,7 +11,7 @@ int helper_random_int_min_max(int i_min, int i_max) {
   return floor(helper_random_float_min_max((float)i_min, (float)i_max));
 }
 
-double helper_random_float_min_max(float f_min, float f_max) {
+float helper_random_float_min_max(float f_min, float f_max) {
   float f = (float)rand() / RAND_MAX;
   return f_min + f * (f_max - f_min);
 }
