+++ helper.h
@@ -1,6 +1,6 @@
 void helper_set_random_seed(void);
 
 int helper_random_int_min_max(int i_min, int i_max);
-double helper_random_float_min_max(float f_min, float f_max);
+float helper_random_float_min_max(float f_min, float f_max);
 
 int helper_random_probability(float p);
